+++ src/main/java/com/shoponline/warehouse/service/WarehouseNotFoundException.java
@@ -2,7 +2,7 @@ package com.shoponline.warehouse.service;
 
 class WarehouseItemNotFoundException extends RuntimeException {
 
-    public WarehouseItemNotFoundException(Long id) {
+    public WarehouseItemNotFoundException(Integer id) {
         super("Could not find item by " + id);
     }
